Simplify InsertP
